+++ packages/amazon-bedrock/src/bedrock-provider.ts
@@ -1,7 +1,7 @@
 import {
   EmbeddingModelV3,
-  ImageModelV2,
-  LanguageModelV2,
+  ImageModelV3,
+  LanguageModelV3,
   ProviderV3,
 } from '@ai-sdk/provider';
 import {
@@ -105,21 +105,21 @@ and `sessionToken` settings.
 }
 
 export interface AmazonBedrockProvider extends ProviderV3 {
-  (modelId: BedrockChatModelId): LanguageModelV2;
+  (modelId: BedrockChatModelId): LanguageModelV3;
 
-  languageModel(modelId: BedrockChatModelId): LanguageModelV2;
+  languageModel(modelId: BedrockChatModelId): LanguageModelV3;
 
   embedding(modelId: BedrockEmbeddingModelId): EmbeddingModelV3<string>;
 
   /**
 Creates a model for image generation.
    */
-  image(modelId: BedrockImageModelId): ImageModelV2;
+  image(modelId: BedrockImageModelId): ImageModelV3;
 
   /**
 Creates a model for image generation.
    */
-  imageModel(modelId: BedrockImageModelId): ImageModelV2;
+  imageModel(modelId: BedrockImageModelId): ImageModelV3;
 
   /**
 Anthropic-specific tools that can be used with Anthropic models on Bedrock.