+++ packages/deepinfra/src/deepinfra-provider.ts
@@ -1,8 +1,8 @@
 import {
-  LanguageModelV2,
+  LanguageModelV3,
   EmbeddingModelV3,
   ProviderV3,
-  ImageModelV2,
+  ImageModelV3,
 } from '@ai-sdk/provider';
 import {
   OpenAICompatibleChatLanguageModel,
@@ -46,32 +46,32 @@ export interface DeepInfraProvider extends ProviderV3 {
   /**
 Creates a model for text generation.
 */
-  (modelId: DeepInfraChatModelId): LanguageModelV2;
+  (modelId: DeepInfraChatModelId): LanguageModelV3;
 
   /**
 Creates a chat model for text generation.
 */
-  chatModel(modelId: DeepInfraChatModelId): LanguageModelV2;
+  chatModel(modelId: DeepInfraChatModelId): LanguageModelV3;
 
   /**
 Creates a model for image generation.
   */
-  image(modelId: DeepInfraImageModelId): ImageModelV2;
+  image(modelId: DeepInfraImageModelId): ImageModelV3;
 
   /**
 Creates a model for image generation.
   */
-  imageModel(modelId: DeepInfraImageModelId): ImageModelV2;
+  imageModel(modelId: DeepInfraImageModelId): ImageModelV3;
 
   /**
 Creates a chat model for text generation.
 */
-  languageModel(modelId: DeepInfraChatModelId): LanguageModelV2;
+  languageModel(modelId: DeepInfraChatModelId): LanguageModelV3;
 
   /**
 Creates a completion model for text generation.
 */
-  completionModel(modelId: DeepInfraCompletionModelId): LanguageModelV2;
+  completionModel(modelId: DeepInfraCompletionModelId): LanguageModelV3;
 
   /**
 Creates a text embedding model for text generation.