+++ packages/google-vertex/src/google-vertex-provider.ts
@@ -1,5 +1,5 @@
 import { GoogleGenerativeAILanguageModel } from '@ai-sdk/google/internal';
-import { ImageModelV2, LanguageModelV2, ProviderV3 } from '@ai-sdk/provider';
+import { ImageModelV3, LanguageModelV3, ProviderV3 } from '@ai-sdk/provider';
 import {
   FetchFunction,
   generateId,
@@ -22,19 +22,19 @@ export interface GoogleVertexProvider extends ProviderV3 {
   /**
 Creates a model for text generation.
    */
-  (modelId: GoogleVertexModelId): LanguageModelV2;
+  (modelId: GoogleVertexModelId): LanguageModelV3;
 
-  languageModel: (modelId: GoogleVertexModelId) => LanguageModelV2;
+  languageModel: (modelId: GoogleVertexModelId) => LanguageModelV3;
 
   /**
    * Creates a model for image generation.
    */
-  image(modelId: GoogleVertexImageModelId): ImageModelV2;
+  image(modelId: GoogleVertexImageModelId): ImageModelV3;
 
   /**
 Creates a model for image generation.
    */
-  imageModel(modelId: GoogleVertexImageModelId): ImageModelV2;
+  imageModel(modelId: GoogleVertexImageModelId): ImageModelV3;
 
   tools: typeof googleVertexTools;
 }