+++ packages/togetherai/src/togetherai-provider.ts
@@ -1,8 +1,8 @@
 import {
-  LanguageModelV2,
+  LanguageModelV3,
   EmbeddingModelV3,
   ProviderV3,
-  ImageModelV2,
+  ImageModelV3,
 } from '@ai-sdk/provider';
 import {
   OpenAICompatibleChatLanguageModel,
@@ -46,22 +46,22 @@ export interface TogetherAIProvider extends ProviderV3 {
   /**
 Creates a model for text generation.
 */
-  (modelId: TogetherAIChatModelId): LanguageModelV2;
+  (modelId: TogetherAIChatModelId): LanguageModelV3;
 
   /**
 Creates a chat model for text generation.
 */
-  chatModel(modelId: TogetherAIChatModelId): LanguageModelV2;
+  chatModel(modelId: TogetherAIChatModelId): LanguageModelV3;
 
   /**
 Creates a chat model for text generation.
 */
-  languageModel(modelId: TogetherAIChatModelId): LanguageModelV2;
+  languageModel(modelId: TogetherAIChatModelId): LanguageModelV3;
 
   /**
 Creates a completion model for text generation.
 */
-  completionModel(modelId: TogetherAICompletionModelId): LanguageModelV2;
+  completionModel(modelId: TogetherAICompletionModelId): LanguageModelV3;
 
   /**
 Creates a text embedding model for text generation.
@@ -73,12 +73,12 @@ Creates a text embedding model for text generation.
   /**
 Creates a model for image generation.
 */
-  image(modelId: TogetherAIImageModelId): ImageModelV2;
+  image(modelId: TogetherAIImageModelId): ImageModelV3;
 
   /**
 Creates a model for image generation.
 */
-  imageModel(modelId: TogetherAIImageModelId): ImageModelV2;
+  imageModel(modelId: TogetherAIImageModelId): ImageModelV3;
 }
 
 export function createTogetherAI(